+++ Thunderstore/CHANGELOG.md
@@ -10,4 +10,8 @@
 - Added Minimum Time configurable property
 - Migrated Benthic-like transformation to its own properties
 - Adjust default weights for a larger difference and a higher focus on void items
-- Nerf default bomb generation frequency (1 -> 0.5)
+- Nerf default bomb generation frequency stacking rate (1 -> 0.5)
+- Buff default bomb damage (200% -> 300%)
+- Nerf default stacking bomb damage (10% -> 5%)
+- Buff default bomb range (15 -> 17.5)
+- Nerf default stacking bomb range (1 -> 0.5)